feat: migrate from Qt5 to Qt6
- Update CMakeLists.txt to use Qt6 instead of Qt5
- Update cmake/*.cmake files for Qt6 compatibility
- Replace deprecated Qt5 APIs with Qt6 equivalents:
  - QRegExp -> QRegularExpression
  - QX11Info -> QNativeInterface::QX11Application
  - QMediaPlayer::State -> QMediaPlayer::PlaybackState
  - QMediaPlayer::stateChanged -> playbackStateChanged
  - setMedia(QMediaContent) -> setSource(QUrl)
  - QAudioInput (recording) -> QAudioSource
  - QAudioDeviceInfo -> QAudioDevice + QMediaDevices
  - QAudioFormat::setSampleSize/setCodec -> setSampleFormat
  - QNetworkRequest::FollowRedirectsAttribute -> RedirectPolicyAttribute
  - Remove Qt::AA_EnableHighDpiScaling (default in Qt6)
  - Remove QTextCodec usage
  - Remove QQuickStyle::availableStyles() (not in Qt6)
- Fix GCC 15 type strictness (std::clamp/max int vs qsizetype)
- Update qhotkey external project to build with Qt6

+++ src/app_server.cpp
@@ -134,7 +134,7 @@ int app_server::request_another_instance(const cmd::options &options) {
                     auto model = qdbus_cast<QVariantMap>(
                         m.template value<QDBusArgument>());
                     return std::max(model.contains("id")
-                                        ? model.value("id").toString().size()
+                                        ? static_cast<int>(model.value("id").toString().size())
                                         : size,
                                     size);
                 });
@@ -204,10 +204,10 @@ int app_server::request_another_instance(const cmd::options &options) {
             g_max_size =
                 std::max(g_max_size,
                          std::max(modelStt.contains("id")
-                                      ? modelStt.value("id").toString().size()
+                                      ? static_cast<int>(modelStt.value("id").toString().size())
                                       : 1,
                                   modelTts.contains("id")
-                                      ? modelTts.value("id").toString().size()
+                                      ? static_cast<int>(modelTts.value("id").toString().size())
                                       : 1));
             print_active_model("STT", g_max_size, modelStt);
             print_active_model("TTS", g_max_size, modelTts);
@@ -215,14 +215,14 @@ int app_server::request_another_instance(const cmd::options &options) {
             auto modelStt = iface.activeSttModel();
             g_max_size = std::max(g_max_size,
                                   modelStt.contains("id")
-                                      ? modelStt.value("id").toString().size()
+                                      ? static_cast<int>(modelStt.value("id").toString().size())
                                       : 1);
             print_active_model("STT", g_max_size, modelStt);
         } else if (options.active_model_to_print_role & cmd::role_tts) {
             auto modelTts = iface.activeTtsModel();
             g_max_size = std::max(g_max_size,
                                   modelTts.contains("id")
-                                      ? modelTts.value("id").toString().size()
+                                      ? static_cast<int>(modelTts.value("id").toString().size())
                                       : 1);
             print_active_model("TTS", g_max_size, modelTts);
         }
